+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.1.0] - 2026-08-06
+### Added
+- Glowing vines (leaked sniffer plant)
+
 ### Changed
 - Sulfur campfire cooked food causes nausea